requirements.txt
# Core dependencies
ScraperFC>=3.2.0
duckdb>=0.10.0
pandas>=2.0.0
polars>=0.20.0
numpy>=1.24.0

# Web scraping
cloudscraper>=1.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# ML dependencies
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0

# Data validation and processing
pydantic>=2.0.0
python-dateutil>=2.8.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.5.0
ipykernel>=6.25.0
jupyter>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
typer>=0.9.0